+++ XGTPro/Utils/PushNotification/PushNotification.h
@@ -15,7 +15,7 @@
 {
 
 }
-@property(nonatomic) int applicationIconBadgeNumber;
+@property(nonatomic) NSInteger applicationIconBadgeNumber;
 @property(nonatomic,assign)id<PushNotificationDelegate> observer;
 +(id)sharePushNotification;
 //根据不同的第三方推送包进行该方法的重写,重写后不调用父类的该方法。
